requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
opencv-python==4.9.0.80